Currently learning how to upload files from VB to Github !
Hey guys .

I am trying to great a program for Dungeon and dragon players. This program will take user input for example : Character name to Skills that charactor has .
Once user has filled the sheet , They can save it to a text file that will be created ( Still working on it ). 
I would really appreciate some help if possible . Its a great learning experience for me to program VB and use GitHub with it .

If you would like to add some stuff please comment and ideas if possible thank you .